api/pyproject.toml
[project]
name = "mon-vrai-projet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "passlib>=1.7.4",
    "pydantic-extra-types>=2.11.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
addopts = "-ra -q --asyncio-mode=auto"
testpaths = ["tests"]
